+++ Services/Interfaces/ITravelExpenseService.cs
@@ -7,7 +7,7 @@ namespace Expense_Tracker.Services.Interfaces
     {
         Task<bool> AddTravelExpense(TravelExpenseDTO travelExpenseDTO);
         Task<bool> RemoveTravelExpense(int travelExpenseId);
-        Task<int> GetTotalTravelExpense(int userId, DateTime date);
+        Task<double> GetTotalTravelExpense(int userId, DateTime date);
         Task<IEnumerable<TravelExpense>> GetAllTravelExpenseOfTheDay(int userId, DateTime date);
         Task<bool> UpdateTravelExpenseById(int TravelExpenseId, TravelExpenseDTO travelExpenseDTO);
     }
